+++ src/components/specSection/TitleText.jsx
